Redefinition of size and type of different component and architecture
- Redefine networks composition (int to short)
- Redefine Gateway composition (add a ID field that will be use in message instead of EUI which is too long)
- Redefine Mote composition (add a ID field that will be use in communication instead of ADDR which is too long)
- Add function for manipulation and correspondence in factory

+++ libraries/ThingSat_v1/factory.h
@@ -12,8 +12,8 @@ float PointToInt(byte point[2]);
 
 /* Gateway things */                     
 boolean IsGatewayDropped(Gateway *gw);                  //Compute if the gateway have to be dropped
-int GetNetworkId(Gateway *gw);                          //return -1 if gw is null
-int GetNetworkId(byte DevADDR[4], RoutingTable *table); //return -1 if table null, -2 if not found
+short GetNetworkId(Gateway *gw);                          //return -1 if gw is null
+short GetNetworkId(byte DevADDR[4], RoutingTable *table); //return -1 if table null, -2 if not found
 boolean IsInSightOf(Gateway *gw);
 Gateway *GetNextInsightGateway();
 boolean AddGateway(Gateway *gw, LPGANNetwork *network, RoutingTable *table);// Return false in case of nullptr or gateway not found, true if it found it and add it
